+++ InterviewAssistant/app/src/main/java/com/example/rafael/interviewassistant/Views/Bhhospital.java
@@ -12,7 +12,7 @@ import com.exemple.rafael.interviewassistant.model.Interview;
 
 public class Bhhospital extends AppCompatActivity {
 
-    private int IdPerson;
+    private Long IdPerson;
     private DataBaseInterview data;
     private String nome;
 
@@ -25,7 +25,7 @@ public class Bhhospital extends AppCompatActivity {
         data = new DataBaseInterview(((App) getApplication()).getDaoSession());
 
         Intent intent = getIntent();
-        IdPerson = intent.getIntExtra("Id",0);
+        IdPerson = intent.getLongExtra("Id", 0);
         nome = intent.getStringExtra("Name");
 
     }
