+++ InterviewAssistant/app/src/main/java/com/example/rafael/interviewassistant/Views/haschildren.java
@@ -12,7 +12,7 @@ import com.exemple.rafael.interviewassistant.model.Interview;
 
 public class haschildren extends ActionBarActivity {
 
-    private int IdPerson;
+    private long IdPerson;
     private DataBaseInterview data;
     private String nome;
 
@@ -22,7 +22,7 @@ public class haschildren extends ActionBarActivity {
         setContentView(R.layout.activity_haschildrens);
 
         Intent intent = getIntent();
-        IdPerson = intent.getIntExtra("Id", 0);
+        IdPerson = intent.getLongExtra("Id", 0);
         nome = intent.getStringExtra("Name");
         data = new DataBaseInterview(((App) getApplication()).getDaoSession());
